feat: added stub classes for query builder and api config

+++ VKFeedStatistics/src/main/kotlin/client/VkClient.kt
@@ -4,5 +4,5 @@ import protocol.VkApiResponse
 import java.io.Closeable
 
 interface VkClient : Closeable {
-    suspend fun countPostByHashtag(hashtag: String, hoursLimit: Int) : VkApiResponse?
+    suspend fun countPostByHashtag(hashtag: String, hoursLimit: Long) : VkApiResponse?
 }
